feat: Day2 CustomNetAdapter（UTP 传输层：连接管理/事件循环/双通道/发送/断开）

+++ Assets/Scripts/Network/INetworkTransport.cs
@@ -7,8 +7,8 @@ namespace Network
         bool IsServer { get; }
         bool IsConnected { get; }
         
-        void StartServer(int port);
-        void ConnectToServer(string ip, int port);
+        void StartServer(ushort port);
+        void ConnectToServer(string ip, ushort port);
         
         void Disconnect();
         void Pump();
